pyproject.toml
[tool.poetry]
name = "devis project"
version = "0.1.0"
description = ""
authors = ["deviyanti_mariam <deviyantiam@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "1.1.1"
numpy = "1.19.2"
Flask = "1.1.2"
requests = "2.25.1"
pickleshare = "0.7.5"
PySastrawi = "1.2.0"
pyspark = "2.4.0"
scikit-learn = "0.20.3"
matplotlib = "3.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"